refactor(config-sync): use useragent with requests
