+++ packages/mc-scripts/src/utils/auth.js
@@ -1,14 +1,5 @@
 const fetch = require('node-fetch');
-const createHttpUserAgent = require('@commercetools/http-user-agent');
-const pkgJson = require('../../package.json');
-
-const userAgent = createHttpUserAgent({
-  name: 'cli-login',
-  libraryName: 'mc-scripts',
-  libraryVersion: pkgJson.version,
-  contactUrl: 'https://git.io/fjuyC', // points to the appkit repo issues
-  contactEmail: 'support@commercetools.com',
-});
+const userAgent = require('./user-agent');
 
 const getAuthToken = async (mcApiUrl, payload) => {
   const response = await fetch(`${mcApiUrl}/tokens/cli`, {
